requirements.txt
flask
pillow
pytesseract
img2pdf
PyMuPDF
